Fixed Ubuntu x86 Clean Bug

+++ modules/clean-packages/ubuntu/post.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 # Post Clean Commands (Ubuntu)
 
-# Clean GCC
-apt-get purge -q -y gcc-4.6-base:i386 libc6:i386 libgcc1:i386
+# Check Architecture
+if [ $(uname -m) = "x86_64" ]; then
+	# Clean GCC
+	apt-get purge -q -y gcc-4.6-base:i386 libc6:i386 libgcc1:i386
+fi
 
 # Clean Python
 echo 'Yes, do as I say!' | apt-get purge -q -y --force-yes python2.7-minimal
